requirements.txt
Django==4.2.1
django-precise-bbcode==1.2.16